+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ v1.0.1 - 2017-08-15
 -  fix smooth option
 -  improve the help output
 -  renamed counter electrode option into --biologic_ce
+-  create journal subcommand
 
 v1.0.0 - 2017-08-13
 -------------------